removing unused old code and moving sensor positions to visualization
