+++ braindecode/datautil/__init__.py
@@ -3,6 +3,5 @@ Utilities for data manipulation.
 """
 
 from .signal_target import SignalAndTarget
-from .loader import CropsDataLoader
 from .windowers import EventWindower, FixedLengthWindower
 from .transforms import FilterRaw, ZscoreRaw, FilterWindow, ZscoreWindow
