+++ braindecode/datautil/iterators.py
@@ -4,105 +4,6 @@ from numpy.random import RandomState
 from braindecode.util import get_balanced_batches
 
 
-class BalancedBatchSizeIterator(object):
-    """
-    Create batches of balanced size.
-    
-    Parameters
-    ----------
-    batch_size: int
-        Resulting batches will not necessarily have the given batch size
-        but rather the next largest batch size that allows to split the set into
-        balanced batches (maximum size difference 1).
-    seed: int
-        Random seed for initialization of `numpy.RandomState` random generator
-        that shuffles the batches.
-    """
-
-    def __init__(self, batch_size, seed=328774):
-        self.batch_size = batch_size
-        self.seed = seed
-        self.rng = RandomState(self.seed)
-
-    def get_batches(self, dataset, shuffle):
-        n_trials = len(dataset.X)
-        batches = get_balanced_batches(
-            n_trials, batch_size=self.batch_size, rng=self.rng, shuffle=shuffle
-        )
-        for batch_inds in batches:
-            batch_X = dataset.X[batch_inds]
-            batch_y = dataset.y[batch_inds]
-
-            # add empty fourth dimension if necessary
-            if batch_X.ndim == 3:
-                batch_X = batch_X[:, :, :, None]
-            yield (batch_X, batch_y)
-
-    def reset_rng(self):
-        self.rng = RandomState(self.seed)
-
-
-class ClassBalancedBatchSizeIterator(object):
-    """
-    Create batches of balanced size, that are also balanced per class, i.e.
-    each class should be sampled roughly with the same frequency during
-    training.
-
-    Parameters
-    ----------
-    batch_size: int
-        Resulting batches will not necessarily have the given batch size
-        but rather the next largest batch size that allows to split the set into
-        balanced batches (maximum size difference 1).
-    seed: int
-        Random seed for initialization of `numpy.RandomState` random generator
-        that shuffles the batches.
-    """
-
-    def __init__(self, batch_size, seed=328774):
-        self.batch_size = batch_size
-        self.seed = seed
-        self.rng = RandomState(self.seed)
-
-    def get_batches(self, dataset, shuffle):
-        n_trials = len(dataset.X)
-        batches = get_balanced_batches(
-            n_trials, batch_size=self.batch_size, rng=self.rng, shuffle=shuffle
-        )
-        if shuffle:
-            n_classes = np.max(dataset.y) + 1
-            class_probabilities = [
-                np.mean(dataset.y == i_class) for i_class in range(n_classes)
-            ]
-            class_probabilities = np.array(class_probabilities)
-            # choose trials in inverse probability of class
-            trial_probabilities = [
-                1.0 / class_probabilities[y] for y in dataset.y
-            ]
-            trial_probabilities = np.array(trial_probabilities) / np.sum(
-                trial_probabilities
-            )
-            i_trial_to_balanced = self.rng.choice(
-                n_trials, n_trials, p=trial_probabilities
-            )
-
-        for batch_inds in batches:
-            if shuffle:
-                batch_inds = [
-                    i_trial_to_balanced[i_trial] for i_trial in batch_inds
-                ]
-            batch_X = dataset.X[batch_inds]
-            batch_y = dataset.y[batch_inds]
-
-            # add empty fourth dimension if necessary
-            if batch_X.ndim == 3:
-                batch_X = batch_X[:, :, :, None]
-            yield (batch_X, batch_y)
-
-    def reset_rng(self):
-        self.rng = RandomState(self.seed)
-
-
 class CropsFromTrialsIterator(object):
     """
     Iterator sampling crops out the trials so that each sample
